+++ CHANGELOG.md
@@ -6,11 +6,18 @@
 
 ***
 
+### Release 19.10
+
+- **xubuntu-vnc-novnc-chromium**
+  - **Chromium Browser** updated to version **77.0.3865.90**
+- **xubuntu-vnc-novnc-firefox**
+  - initial image with **Firefox Quantum** version **69.0.1**
+
 ### Release 19.09
 
-- **xubuntu-vnc**
+- **xubuntu-vnc-novnc**
   - initial image with **Ubuntu 18.04.3 LTS**
-- **xubuntu-vnc-chromium**
+- **xubuntu-vnc-novnc-chromium**
   - initial image with **Chromium Browser** version **76.0.3809.100**
 
 ***